// projects/emr-application/src/app/modules/patient/components/medical.note/components/plan/model/Specialties.ts
export interface Specialties {
  modalitiesOrthoticFabrication: boolean;
  modalitiesOrthoticFabricationNotes: string;
  modalitiesTensFitting: boolean;
  modalitiesTensFittingNotes: string;
  modalitiesAcupuncture: boolean;
  modalitiesAcupunctureNotes: string;
  modalitiesOther: boolean;
  modalitiesOtherNotes: string;
}